refactor(ffmpeg): split normalize.rs into submodules (#42)
* refactor(ffmpeg): deduplicate normalize.rs — extract helpers, unify encode/dispatch

Extract 5 shared patterns to eliminate ~95% code duplication between
peak and loudnorm normalization paths:

- drain_astats_metadata: replaces 3 identical metadata-drain loops in
  analyze_peak_sync
- last_stderr_lines: replaces 2 identical stderr extraction patterns in
  CLI fallback functions
- cli_fallback_normalize: unifies cli_fallback_peak and
  cli_fallback_loudnorm into a single function parameterized by filter
  string and label
- encode_audio_only_sync: unifies peak_encode_audio_only and
  loudnorm_encode_audio_only (~420 lines each, 95% identical) into a
  single function with a build_filter closure for the only difference
- dispatch_normalize_sync: unifies apply_peak_gain_sync and
  loudnorm_pass2_sync dispatch logic (has_video check, temp audio path,
  salvage retry, merge, cleanup)

2651 → 2186 lines (465 lines removed, ~17.5% reduction).
No behavioral changes — all 74 tests pass.

* refactor(ffmpeg): split normalize.rs into submodules, extract common analysis loop

Split the 2,186-line normalize.rs into a directory module with 6 focused
files (all ≤526 LOC). Extract run_analysis_decode_loop() to deduplicate
the ~80% shared pipeline between analyze_peak_sync and loudnorm_pass1_sync.

- normalize/mod.rs: public API, dispatch, limiter-boost
- normalize/analysis.rs: shared decode loop, peak/loudnorm analysis
- normalize/encode.rs: unified encode pipeline, mode wrappers
- normalize/helpers.rs: filter builders, JSON parsing, CLI fallbacks
- normalize/io_diag.rs: AVIO diagnostics, seekability validation
- normalize/tests.rs: all 20 existing tests (unchanged)
