+++ crates/rdlp-ffmpeg/src/ffmpeg/merge.rs
@@ -235,8 +235,7 @@ impl FFmpegRunner {
         Self::clear_codec_tag(ost_audio.parameters().as_ptr());
         // Set audio as default stream so players select it automatically
         unsafe {
-            (*ost_audio.as_mut_ptr()).disposition =
-                ffmpeg_the_third::ffi::AV_DISPOSITION_DEFAULT;
+            (*ost_audio.as_mut_ptr()).disposition = ffmpeg_the_third::ffi::AV_DISPOSITION_DEFAULT;
         }
         let audio_ost_index = ost_audio.index();
 
@@ -254,9 +253,7 @@ impl FFmpegRunner {
                 message: format!("failed to write output header: {e}"),
             })?;
 
-        info!(
-            "Merge: video=stream#{video_ost_index}, audio=stream#{audio_ost_index} (DEFAULT)"
-        );
+        info!("Merge: video=stream#{video_ost_index}, audio=stream#{audio_ost_index} (DEFAULT)");
 
         // Two-way timestamp-interleaved merge: read packets from both inputs
         // and write them in DTS order to avoid ENOMEM from buffering an entire
@@ -657,9 +654,7 @@ impl FFmpegRunner {
                 });
             }
 
-            info!(
-                "Merge: video=stream#{video_out_idx}, audio=stream#{audio_out_idx} (DEFAULT)"
-            );
+            info!("Merge: video=stream#{video_out_idx}, audio=stream#{audio_out_idx} (DEFAULT)");
 
             // 13. Two-way timestamp-interleaved merge
             //
